requirements.txt
requests
PyYAML
python-dateutil
feedparser
pandas
openai
tqdm
beautifulsoup4
google-generativeai
flask
gspread
google-auth
PyGithub
